% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMean.R
\name{myMean}
\alias{myMean}
\title{myMean()}
\usage{
myMean(assignment2)
}
\arguments{
\item{assignment2}{numeric vector}
}
\value{
numeric mean
}
\description{
myMean()
}
\examples{
assignment2 <- c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22)
myMean(assignment2)
}
